Minimum-voltage-based balancing algorithm
Wrote a balancing algorithm that discharges all the cells that are
greater than the minimum cell voltage by more than the specified
tolerance. This is the algorithm I found in several lithium BMS papers,
and should ultimately do the same thing as Lachlan's
average-voltage-based algorithm, but a bit simpler and faster.